Pull Tokenizer out of GlslParser.
